fix: startIndex variable renamed to index  and comment out unnecessary code.
This commit fixes a bug in the `document` function where the variable `index` was not correctly assigned from the `data.filter` object. Instead of using `data.filter.startIndex`, it should use `data.filter.index`. The code has been updated to use the correct variable. Additionally, the unnecessary code that slices the `documents` array based on `index` and `limit` values has been commented out since it is not currently needed.